Function lie::root_system::find_roots_from_simple

pub fn find_roots_from_simple(simple_roots: &[Root]) -> Vec<Root>

Find all the roots given a set of simple roots.

This algorithmically finds and creates all of the roots in a root system given a set of simple roots. The resulting list is sorted by level.

This algorithm can be quite time consuming, especially for very large root systems (more than 20 simple roots), but will automatically multithread in such circumstances. Ideally, this list should be generated once and then stored for later use as needed.

Example

use lie::Root;
use lie::root_system::find_roots_from_simple;

let simple_roots = vec![
    Root::simple(vec![2, -1], 0),
    Root::simple(vec![-3, 2], 1),
];
let roots = find_roots_from_simple(&simple_roots);
assert_eq!(roots.len(), 14);

Panics

Panics if the list of simple roots is empty. Also panics if any of the simple roots provided is actually not simple.