Function lie::root_system::find_simple_roots

pub fn find_simple_roots(cm: &CartanMatrix) -> Vec<Root>

Find all the simple roots from the Cartan matrix.

Example

use lie::{Root, CartanMatrix};
use lie::root_system::find_simple_roots;

let cm = CartanMatrix::from(vec![[2, -1], [-3, 2]]);
let simple_roots = find_simple_roots(&cm);
assert_eq!(simple_roots.len(), 2);

Panics

Panics if the Cartan matrix is not square, or if it has size 0.