Trait lie::root_system::RootSystem

pub trait RootSystem {
    fn cartan_matrix(&self) -> &CartanMatrix;

    fn inverse_cartan_matrix(&self) -> &InverseCartanMatrix;

    fn basis_lengths(&self) -> &BasisLengths;

    fn simple_roots(&self) -> &[Root];

    fn positive_roots(&self) -> &[Root];

    fn roots(&self) -> &[Root];

    fn rank(&self) -> usize { ... }

    fn determinant(&self) -> i64 { ... }

    fn num_simple_roots(&self) -> usize { ... }

    fn num_positive_roots(&self) -> usize { ... }

    fn num_roots(&self) -> usize { ... }

    fn inner_product<'a, 'b>(&'a self, a: &'b Root, b: &'b Root) -> i64 { ... }

    fn scalar_product(&self, b: &Root, a: &Root) -> i64 { ... }
}

Trait for root systems.

Definition

Let \(V\) be a finite dimensional Euclidean vector space with the standard Euclidean inner product being denoted by \(\langle \cdot, \cdot \rangle\). A root system in \(V\) is a finite set \(\Phi\) of non-zero vectors (called roots) that satisfy the following conditions:

The roots span \(V\);

The only scalar multiple of a root \(\alpha \in \Phi\) that belong to \(\Phi\) are \(\alpha\) and \(-\alpha\);

For every root \(\alpha \in \Phi\), the set \(\Phi\) is closed under reflections through the hyperplane perpendicular to \(\alpha\);

If \(\alpha, \beta \in \Phi\), then the projection of \(\beta\) onto \(\alpha\) is a half-integral multiple of \(\alpha\).

In addition to the set of roots themselves, there are two important subsets of \(\Phi\) which are commonly used:

\(\Phi^{+} \subseteq \Phi\): the set of positive roots. This is the subset chosen such that:

For each \(\alpha \in \Phi\), only one of \(\alpha\) and \(-\alpha\) are in \(\Phi^{+}\); and,

For any two distinct \(\alpha, \beta \in \Phi^{+}\) then \(\alpha + \beta \in \Phi^{+}\) provided that \(\alpha + \beta \in \Phi\).

\(\Delta \subseteq \Phi^{+}\): the set of simple root. An element of \(\Phi^{+}\) is a simple root if it cannot be expressed as the sum of two other elements in \(\Phi^{+}\). The set of simple roots has the property that every root in \(\Phi\) can be expressed as linear combination of elements of \(\Delta\) with all coefficients non-negative, or all coefficients non-positive.

Although the roots are, strictly speaking, elements of a Euclidean vector space; we will not be handling these vectors directly. Instead, all the relevant information is contained in the [CartanMatrix].

Required Methods

fn cartan_matrix(&self) -> &CartanMatrix

Return the Cartan matrix of the Lie group.

Example

The Cartan matrix for \(\mathrm{SU}(3)\) is:

\begin{equation} \begin{pmatrix} 2 & -1 \\ -1 & 2 \end{pmatrix} \end{equation}

fn inverse_cartan_matrix(&self) -> &InverseCartanMatrix

Return the inverse Cartan matrix of the Lie group.

See [InverseCartanMatrix] for details as to how the matrix is stored within the library.

fn basis_lengths(&self) -> &BasisLengths

Return the squared norm of each simple roots.

Example

The squared norms of the three simple roots in \(B_{3} = \mathrm{SO}(7)\) are:

\begin{equation} \left( 2, 2, 1 \right) \end{equation}

The three roots in Euclidean space are:

\begin{align} \alpha_{1} &= (1, -1, 0) \\ \alpha_{2} &= (0, 1, -1) \\ \alpha_{3} &= (0, 0, 1) \end{align}

fn simple_roots(&self) -> &[Root]

Return the simple roots of the Lie group's root system.

The roots are sorted such that they correspond with the Cartan matrix.

This can be algorithmically created using the find_simple_roots function.

Definition

The set of simple roots, \(\Delta\), consists of the subset of positive roots which cannot be expressed as the sum of other elements in \(\Phi^{+}\).

Example

The simple roots of \(\mathrm{SU}(3)\) are:

\begin{align} (-1, 2) &~ (2, -1) \\ \end{align}

fn positive_roots(&self) -> &[Root]

Return the positive roots of the Lie group's root system.

The roots should be ordered by level, but no particular ordering should be assumed within each level.

This can be algorithmically created using the find_positive_roots function.

Definition

The set of positive roots, \(\Phi^{+} \subseteq \Phi\), is the subset chosen such that:

For each \(\alpha \in \Phi\), only one of \(\alpha\) and \(-\alpha\) are in \(\Phi^{+}\); and,

For any two distinct \(\alpha, \beta \in \Phi^{+}\) then \(\alpha+\beta \in \Phi^{+}\) provided that \(\alpha + \beta \in \Phi\).

Example

The positive roots of \(\mathrm{SU}(3)\) are:

\begin{align} (1, &~ 1) \\ (-1, 2) &~ (2, -1) \\ \end{align}

fn roots(&self) -> &[Root]

Return the roots in the Lie group's root system.

The roots should be ordered by level, but no particular ordering should be assumed within each level.

This can be algorithmically created using the find_roots_from_simple or find_roots_from_positive functions.

Definition

The roots in the Lie group's root system corresponds to the number of generators for the Lie group.

Example

The Lie group \(\mathrm{SU}(3)\) has 8 generators (one common choice of generators is the 8 Gell-Mann matrices), and thus has 8 roots:

\begin{align} (1, &~ 1) \\ (-1, 2) &~ (2, -1) \\ (0, 0) &~ (0, 0) \\ (-2, 1) &~ (1, -2) \\ (-1, &~ {-1}) \end{align}

Provided Methods

fn rank(&self) -> usize

Return the rank of the Lie group.

Definition

The rank of a Lie group is the dimensionality of the group's Cartan subgroup. In the case of groups of matrices, the Cartan subgroup corresponds to the subgroup of diagonal matrices.

The rank of the Lie group also corresponds to the number of simple roots in the Lie group's root system.

Example

The Lie group \(\mathrm{SU}(3)\) has rank 2 and the two generators of the Cartan subgroup are:

\begin{equation} \left\{ \begin{pmatrix} 1 & 0 & 0 \\ 0 & -1 & 0 \\ 0 & 0 & 0 \end{pmatrix} , \frac{1}{\sqrt{3}} \begin{pmatrix} 1 & 0 & 0 \\ 0 & 1 & 0 \\ 0 & 0 & -2 \end{pmatrix} \right\} \end{equation}

In the case of \(\mathrm{SU}(3)\), the simple roots are \(\{(2, -1), (-1, 2)\}\) and correspond to the rows of the Cartan matrix of \(\mathrm{SU}(3)\).

fn determinant(&self) -> i64

Return the determinant of the Cartan matrix of the Lie group.

fn num_simple_roots(&self) -> usize

Return the number of simple roots in the Lie group's root system.

Warning

The default implementation of this function will generate all the simple roots in the root system, and then return the length of the corresponding vector. If there exists a simple closed form for the number of simple roots, it should be implemented.

fn num_positive_roots(&self) -> usize

Return the number of positive roots in the Lie group's root system.

fn num_roots(&self) -> usize

Return a number of roots in the Lie group's root system.

fn inner_product<'a, 'b>(&'a self, a: &'b Root, b: &'b Root) -> i64

Evaluate the Euclidean inner product between two roots of the system.

The Euclidean product, denoted by \(( \cdot, \cdot)\), is the usual inner product from Euclidean space.

The roots are scaled such that the product is always an integer. See [BasisLengths] for more information.

Example

The roots of \(B_{3}\) in Euclidean space are

\begin{align} \alpha_{1} &= (1, -1, 0) \\ \alpha_{2} &= (0, 1, -1) \\ \alpha_{3} &= (0, 0, 1) \end{align}

So for example \(( \alpha_{1}, \alpha_{2} ) = 2\), and \(( \alpha_{1}, \alpha_{2} ) = -1\).

Panics

Panics if the roots do not have the same rank as the root system.

fn scalar_product(&self, b: &Root, a: &Root) -> i64

Evaluate the scalar product between two roots of the system.

The scalar product between roots is defined as

\begin{equation} \langle \beta, \alpha \rangle \equiv 2 \frac{(\alpha, \beta)}{(\alpha, \alpha)} \end{equation}

where \((\cdot, \cdot)\) is the usual Euclidean inner product between the roots in Euclidean space. This should always evaluate to an integer in a root system.

Note that this scalar product is only linear in its first argument. It is not an inner product.