Type Definition lie::root_system::BasisLengths

type BasisLengths = Array1<i64>;

Type definition for a basis lengths.

Definition

The array is defined as

\begin{equation} D_{i} = (\alpha_{i}, \alpha_{i}) \end{equation}

and must use the same labelling convention for \(\alpha_{i}\) as used in the Cartan matrix. Although the roots can be arbitrarily rescaled by a common factor, they should be chosen such that each \(D_{i}\) is an integer and such that the inner product below always produces an integer.

In combination with the Cartan matrix, this allows for the inner product of any two simple roots to be evaluated,

\begin{equation} ( \alpha_{i}, \alpha_{j} ) = \frac{1}{2} A_{ji} D_{i}, \end{equation}

which can then be further generalized for any two roots \(\beta_{n} = \sum_{i} a_{ni} \alpha_{i}\):

\begin{equation} ( \beta_{n}, \beta_{m} ) = \sum_{ij} a_{ni} b_{mj} ( \alpha_{i}, \alpha_{j} ). \end{equation}

Note that in order to simplify the numerical computations, all roots should be scaled such that \(\langle \alpha_{i}, \alpha_{j} \rangle\) returns an integer.

This array is closely related to the \(D_{ij}\) matrix defined by

\begin{equation} D_{ij} = \frac{\delta_{ij}}{(\alpha_{i}, \alpha_{i})} \end{equation}

but it isn't used here in order to avoid requiring the use of rational numbers.