Module lie::series

Definitions for the 7 families of Lie groups.

Four of the families form infinite series:

\(A_{n} = \mathrm{SU}(n+1)\), the special unitary group;

\(B_{n} = \mathrm{SO}(2n + 1)\), the special orthogonal group of odd dimension;

\(C_{n} = \mathrm{Sp}(2n)\), the group of unitary symplectic matrices; and,

\(D_{n} = \mathrm{SO}(2n)\), the special orthogonal group of even dimension.

The last three families contain the 5 exceptional Lie groups:

\(G_{2}\);

\(F_{4}\);

\(E_{6}\);

\(E_{7}\); and,

\(E_{8}\).

The corresponding Lie algebras are conventionally denoted using fraktur symbols: \(\mathfrak{su}_{n}\), \(\mathfrak{so}_{n}\), \(\mathfrak{sp}_{n}\), \(\mathfrak{g}_{2}\), \(\mathfrak{f}_{4}\), \(\mathfrak{e}_{6}\), \(\mathfrak{e}_{7}\) and \(\mathfrak{e}_{8}\).

Exceptional Isomorphisms

For low values of \(n\), there are accidental isomorphisms between the above categories. In particular, they are:

\(A_{1} \cong B_{1} \cong C_{1}\),

\(B_{2} \cong C_{2}\),

\(D_{2} \cong A_{1} \times A_{1}\),

\(D_{3} \cong A_{3}\),

\(E_{3} \cong A_{1} \times A_{2}\)

\(E_{4} \cong A_{4}\),

\(E_{5} \cong D_{5}\).

As a result, the indices for \(A_{n}\), \(B_{n}\), \(C_{n}\) and \(D_{n}\) start at 1, 2, 3 and 4 respectively.

This library does not handle exceptional isomorphisms, though it might return an error indicating what the standard label for the specified group is. For example, trying to create an instance of the \(B_{1} = \mathrm{SO}(3)\) group will produce an error and instead one must use the isomorphic group \(A_{1} = \mathrm{SU}(2)\).

Structs

TypeA
The \(A_{n}\) infinite series of Lie groups.

TypeB
The \(B_{n}\) infinite series of Lie groups.

TypeC
The \(C_{n}\) infinite series of Lie groups.

TypeD
The \(D_{n}\) infinite series of Lie groups.

TypeE
The \(E_{n}\) exceptional Lie groups.

TypeF
The \(F_{n}\) exceptional Lie groups.

TypeG
The \(G_{n}\) exceptional Lie groups.