Struct lie::Root

pub struct Root { /* fields omitted */ }

Instance of a root.

A root is a vector in the underlying (Euclidean) vector space. Note that we do not store these vectors from the underlying vector space, but instead use a different basis which is dependent on the root system and the selection of simple roots.

The roots are stored in the following two bases:

\(\alpha\)-basis: Represents the root \(\beta\) as the array \((k_{1}, \dots, k_{n})\) such that \(\beta = \sum_{i} k_{i} \alpha_{i}\), where \(\alpha_{i}\) are the simple roots. Note that \(n\) corresponds to the number of simple roots and thus the rank of the corresponding Lie group.

\(\omega\)-basis: In the \(\omega\) basis, each \(\alpha_{i}\) corresponds to the \(i\)th row of Cartan matrix; thus, a root \(\beta\) in the \(\omega\) basis is expressed \(\beta = \sum_{i} k_{i} A_{ij}\).

A root on its own makes little sense without the corresponding [RootSystem]; however, it is still possible to manipulate roots in several ways without needing the additional information provided by the root system. For this to be possible though, it assumes that the \(\omega\) and \(\alpha\) weights have been correctly specified. Incorrectly defining \(\omega\) or \(\alpha\) may result in undefined behaviour when roots are being manipulated.

Methods

impl Root

pub fn zero(rank: usize) -> Self

Create a zero-root.

A zero root has all weights set to zero and is at level 0.

Example

use lie::Root;

let r = Root::zero(5);
assert_eq!(r.level(), 0);
assert_eq!(r.rank(), 5);
assert!(!r.is_simple());

Panics

Panics if the rank is 0.

pub fn simple<A>(omega: A, alpha: usize) -> Self where
A: Into<Array1<i64>>,

Create a simple root.

A simple root is a root which cannot be written as the linear combination of any other root in the system. Note that creating a simple root through this function defines the root to be simple, and it will depend on the root system itself whether this is true or not.

Example

use lie::Root;

let r = Root::simple(vec![2, -1, 0], 0);
assert!(r.is_simple());
assert_eq!(r.level(), 1);
assert_eq!(r.alpha().as_slice().unwrap(), &[1, 0, 0]);

Panics

Panics if alpha is greater than the length of the omega basis (or equivalently, if alpha is greater than the rank of the corresponding root system).

pub fn new<A, B>(omega: A, alpha: B) -> Self where
A: Into<Array1<i64>>,
B: Into<Array1<i64>>,

Create a new arbitrary root with all coefficients in the \(\omega\) and \(\alpha\) bases specified.

This function should not be used unless you are sure of what you are doing. Incorrectly or inconsistently defining omega with respect to alpha (or simply on their own) may have serious repercussions when the roots are being manipulated with other roots.

Example

use lie::Root;

let r = Root::new(vec![2, -1, 0], vec![1, 0, 0]);
assert!(r.is_simple());
assert_eq!(r.level(), 1);

let r = Root::new(vec![2, -1, 0], vec![1, 0, 0]);
assert!(r.is_simple());
assert_eq!(r.level(), 1);

Panics

Panics if the sizes of omega and alpha differ or if they are zero. Also panics if the coefficients in alpha don't all have the same sign.

pub fn rank(&self) -> usize

Return the rank of the root's corresponding Lie group.

Although the root on its own has no way of this, the number of coefficient in the root's \(\alpha\)-basis is the number of simple roots of the root system which corresponds to the Lie group's rank.

Example

use lie::Root;

let r = Root::simple(vec![2, -1, 0], 0);
assert_eq!(r.rank(), 3);

pub fn level(&self) -> i64

Return the level of the root.

A root \(\beta\) can be written uniquely as a linear combination of simple roots \(\alpha_{i}\), \(\beta = \sum_{i} k_{i} \alpha_{i}\). The level of \(\beta\) is then defined as \(\sum_{i} k_{i}\).

Example

use lie::Root;

let a1 = Root::simple(vec![2, -1, 0], 0);
let a2 = Root::simple(vec![-1, 2, -1], 1);

assert_eq!(a1.level(), 1);
assert_eq!((&a1 + &a2).level(), 2);
assert_eq!((&a1 + &a2 + &a2).level(), 3);

pub fn omega(&self) -> &Array1<i64>

Return the root's decomposition in terms of the \(\omega\) basis.

Example

use lie::Root;

let a1 = Root::simple(vec![2, -1, 0], 0);
let a2 = Root::simple(vec![-1, 2, -1], 1);

let r = &a1 + &a2 + &a2;
assert_eq!(a1.omega().as_slice().unwrap(), &[2, -1, 0]);
assert_eq!(a2.omega().as_slice().unwrap(), &[-1, 2, -1]);
assert_eq!(r.omega().as_slice().unwrap(), &[0, 3, -2]);

pub fn alpha(&self) -> &Array1<i64>

Return the root's decomposition in terms of the \(\alpha\) basis.

Example

use lie::Root;

let a1 = Root::simple(vec![2, -1, 0], 0);
let a2 = Root::simple(vec![-1, 2, -1], 1);

let r = &a1 + &a2 + &a2;
assert_eq!(r.level(), 3);
assert_eq!(r.alpha().as_slice().unwrap(), &[1, 2, 0]);

pub fn is_simple(&self) -> bool

Check whether the root is a simple root or not.

Note that this ultimately requires information about the root system, which the root does not have access to. As a result, this function's accuracy relies on the root having been correctly generated in the first place.

Example

use lie::Root;

let r = Root::zero(5);
assert!(!r.is_simple());

let r = Root::simple(vec![2, -1, 0], 0);
assert!(r.is_simple());